# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for a multi-tenant (SaaS) ophthalmology clinic\nmanagement system: patients, visits, billing/POS, inventory,\nlab jobs, reports, and optical calculators.\n\nEvery route under /api (other than /api/auth/*, /api/signup,\nand the /api/auth/password-reset|email endpoints) requires a\nBearer access token, and most additionally require the\ncaller's clinic to have an active subscription.",
        "title": "Opthalmic Management System API",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api",
    "paths": {
...
        "/reports/dues": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Add ?format=csv to download instead of receiving JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Outstanding patient balances",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PatientDueResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reports/inventory/low-stock": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Inventory low-stock report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.InventoryItemResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reports/inventory/valuation": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Inventory valuation report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.InventoryValuationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reports/sales/daily": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Add ?format=csv to download instead of receiving JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Daily sales report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD, defaults to today",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SalesSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reports/sales/monthly": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Add ?format=csv to download instead of receiving JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Monthly sales report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Defaults to current year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Defaults to current month",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SalesSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reports/sales/range": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Add ?format=csv to download instead of receiving JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Sales report over a date range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SalesSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reports/vendor-dues": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Add ?format=csv to download instead of receiving JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Outstanding vendor balances",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.VendorDueResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reports/visits/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Add ?format=csv to download instead of receiving JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Visits summary report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "'day' (default) or 'doctor'",
                        "name": "group_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.VisitsSummaryResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.InventoryItemResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "cost_price": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quantity_on_hand": {
                    "type": "integer"
                },
                "reorder_threshold": {
                    "type": "integer"
                },
                "selling_price": {
                    "type": "number"
                },
                "size": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.InventoryValuationResponse": {
            "type": "object",
            "properties": {
                "total_cost_value": {
                    "type": "number"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_sell_value": {
                    "type": "number"
                },
                "total_units": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PatientDueResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "due_amount": {
                    "type": "number"
                },
                "invoice_no": {
                    "type": "string"
                },
                "paid_amount": {
                    "type": "number"
                },
                "patient_name": {
                    "type": "string"
                },
                "patient_phone": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "dto.SalesSummaryResponse": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "number"
                },
                "invoice_count": {
                    "type": "integer"
                },
                "period": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                },
                "tax": {
                    "type": "number"
                },
                "total_due": {
                    "type": "number"
                },
                "total_paid": {
                    "type": "number"
                },
                "total_sales": {
                    "type": "number"
                }
            }
        },
...
        "dto.VendorDueResponse": {
            "type": "object",
            "properties": {
                "due_amount": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "purchase_date": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "vendor_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.VisitsSummaryResponse": {
            "type": "object",
            "properties": {
                "period": {
                    "type": "string"
                },
                "visit_count": {
                    "type": "integer"
                }
            }
        },
...
        "http.response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token, e.g. \"Bearer eyJhbGci...\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}